requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.21
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==10.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.13.2
  